EPMLABSBRN-182 semantic reorganize code
