+++ src/main/kotlin/com/epam/brn/auth/AuthenticationService.kt
@@ -1,4 +1,4 @@
-package com.epam.brn.service
+package com.epam.brn.auth
 
 import com.epam.brn.dto.LoginDto
 import com.epam.brn.dto.UserAccountDto
