+++ src/main/kotlin/com/epam/brn/auth/AuthorityService.kt
@@ -1,4 +1,4 @@
-package com.epam.brn.service
+package com.epam.brn.auth
 
 import com.epam.brn.model.Authority
