+++ src/main/kotlin/com/epam/brn/auth/AuthorityServiceImpl.kt
@@ -1,9 +1,8 @@
-package com.epam.brn.service.impl
+package com.epam.brn.auth
 
 import com.epam.brn.exception.EntityNotFoundException
 import com.epam.brn.model.Authority
 import com.epam.brn.repo.AuthorityRepository
-import com.epam.brn.service.AuthorityService
 import org.apache.logging.log4j.kotlin.logger
 import org.springframework.stereotype.Service
