+++ src/main/kotlin/com/epam/brn/cloud/CloudService.kt
@@ -1,4 +1,4 @@
-package com.epam.brn.service
+package com.epam.brn.cloud
 
 interface CloudService {
     /**
