+++ src/main/kotlin/com/epam/brn/controller/AuthenticationController.kt
@@ -1,9 +1,9 @@
 package com.epam.brn.controller
 
+import com.epam.brn.auth.AuthenticationService
 import com.epam.brn.dto.AuthOutDto
 import com.epam.brn.dto.LoginDto
 import com.epam.brn.dto.UserAccountDto
-import com.epam.brn.service.AuthenticationService
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import org.springframework.http.HttpStatus
