+++ src/main/kotlin/com/epam/brn/exception/EntityNotFoundException.kt
@@ -1,5 +1,3 @@
 package com.epam.brn.exception
 
-import java.lang.RuntimeException
-
 class EntityNotFoundException(message: String) : RuntimeException(message)
