+++ src/main/kotlin/com/epam/brn/exception/FileFormatException.kt
@@ -1,5 +1,3 @@
 package com.epam.brn.exception
 
-import java.lang.RuntimeException
-
 class FileFormatException : RuntimeException("Formatting error. Please upload file with csv extension.")
