+++ src/main/kotlin/com/epam/brn/localization/LocalePostprocessor.kt
@@ -1,4 +1,4 @@
-package com.epam.brn.service
+package com.epam.brn.localization
 
 interface LocalePostprocessor<Dto> {
     fun postprocess(dto: Dto): Dto
