+++ src/main/kotlin/com/epam/brn/repo/ExerciseRepository.kt
@@ -1,9 +1,9 @@
 package com.epam.brn.repo
 
 import com.epam.brn.model.Exercise
-import java.util.Optional
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
+import java.util.Optional
 
 @Repository
 interface ExerciseRepository : JpaRepository<Exercise, Long> {
