+++ src/main/kotlin/com/epam/brn/repo/ResourceRepository.kt
@@ -1,9 +1,9 @@
 package com.epam.brn.repo
 
 import com.epam.brn.model.Resource
-import java.util.Optional
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
+import java.util.Optional
 
 @Repository
 interface ResourceRepository : CrudRepository<Resource, Long> {
