+++ src/main/kotlin/com/epam/brn/repo/SeriesRepository.kt
@@ -1,10 +1,10 @@
 package com.epam.brn.repo
 
 import com.epam.brn.model.Series
-import java.util.Optional
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
+import java.util.Optional
 
 @Repository
 interface SeriesRepository : CrudRepository<Series, Long> {
