+++ src/main/kotlin/com/epam/brn/repo/StudyHistoryRepository.kt
@@ -1,11 +1,11 @@
 package com.epam.brn.repo
 
 import com.epam.brn.model.StudyHistory
-import java.util.Optional
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.CrudRepository
 import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
+import java.util.Optional
 
 @Repository
 interface StudyHistoryRepository : CrudRepository<StudyHistory, Long> {
