+++ src/main/kotlin/com/epam/brn/repo/TaskRepository.kt
@@ -1,10 +1,10 @@
 package com.epam.brn.repo
 
 import com.epam.brn.model.Task
-import java.util.Optional
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
+import java.util.Optional
 
 @Repository
 interface TaskRepository : JpaRepository<Task, Long> {
