+++ src/main/kotlin/com/epam/brn/repo/UserAccountRepository.kt
@@ -1,10 +1,10 @@
 package com.epam.brn.repo
 
 import com.epam.brn.model.UserAccount
-import java.util.Optional
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
+import java.util.Optional
 
 @Repository
 interface UserAccountRepository : JpaRepository<UserAccount, Long> {
