+++ src/main/kotlin/com/epam/brn/service/impl/UserAccountServiceImpl.kt
@@ -1,13 +1,12 @@
 package com.epam.brn.service.impl
 
+import com.epam.brn.auth.AuthorityService
 import com.epam.brn.dto.UserAccountDto
 import com.epam.brn.exception.EntityNotFoundException
 import com.epam.brn.model.Authority
 import com.epam.brn.model.UserAccount
 import com.epam.brn.repo.UserAccountRepository
-import com.epam.brn.service.AuthorityService
 import com.epam.brn.service.UserAccountService
-import java.security.Principal
 import org.apache.commons.lang3.StringUtils.isNotEmpty
 import org.apache.logging.log4j.kotlin.logger
 import org.springframework.security.core.Authentication
@@ -16,6 +15,7 @@ import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UsernameNotFoundException
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
+import java.security.Principal
 
 @Service
 class UserAccountServiceImpl(
