+++ src/main/kotlin/com/epam/brn/upload/csv/series2/SeriesTwoRecordMappingIteratorProvider.kt
@@ -4,8 +4,8 @@ import com.epam.brn.upload.csv.MappingIteratorProvider
 import com.fasterxml.jackson.databind.MappingIterator
 import com.fasterxml.jackson.dataformat.csv.CsvMapper
 import com.fasterxml.jackson.dataformat.csv.CsvParser
-import java.io.InputStream
 import org.springframework.stereotype.Component
+import java.io.InputStream
 
 @Component
 class SeriesTwoRecordMappingIteratorProvider :
