+++ src/test/kotlin/com/epam/brn/controller/GroupControllerTest.kt
@@ -1,10 +1,9 @@
 package com.epam.brn.controller
 
 import com.epam.brn.dto.ExerciseGroupDto
+import com.epam.brn.localization.LocalePostprocessor
 import com.epam.brn.service.ExerciseGroupsService
-import com.epam.brn.service.LocalePostprocessor
 import com.nhaarman.mockito_kotlin.verify
-import kotlin.test.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -12,6 +11,7 @@ import org.mockito.InjectMocks
 import org.mockito.Mock
 import org.mockito.Mockito
 import org.mockito.junit.jupiter.MockitoExtension
+import kotlin.test.assertEquals
 
 @ExtendWith(MockitoExtension::class)
 internal class GroupControllerTest {
