+++ src/test/kotlin/com/epam/brn/service/ExerciseGroupServiceTest.kt
@@ -3,7 +3,6 @@ package com.epam.brn.service
 import com.epam.brn.dto.ExerciseGroupDto
 import com.epam.brn.model.ExerciseGroup
 import com.epam.brn.repo.ExerciseGroupRepository
-import java.util.Optional
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
@@ -14,6 +13,7 @@ import org.mockito.Mockito.`when`
 import org.mockito.Mockito.anyLong
 import org.mockito.Mockito.mock
 import org.mockito.junit.jupiter.MockitoExtension
+import java.util.Optional
 
 @ExtendWith(MockitoExtension::class)
 internal class ExerciseGroupServiceTest {
