+++ src/test/kotlin/com/epam/brn/service/impl/AwsCloudServiceTest.kt
@@ -4,6 +4,7 @@ import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.ListObjectsV2Request
 import com.amazonaws.services.s3.model.ListObjectsV2Result
 import com.amazonaws.services.s3.model.S3ObjectSummary
+import com.epam.brn.cloud.AwsCloudService
 import com.epam.brn.config.AwsConfig
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertEquals
