+++ src/test/kotlin/com/epam/brn/upload/CsvUploadServiceTest.kt
@@ -2,7 +2,6 @@ package com.epam.brn.upload
 
 import com.epam.brn.upload.csv.CsvParser
 import com.epam.brn.upload.csv.RecordProcessor
-import java.io.IOException
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
@@ -10,6 +9,7 @@ import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.InjectMocks
 import org.mockito.Mock
 import org.mockito.junit.jupiter.MockitoExtension
+import java.io.IOException
 
 @ExtendWith(MockitoExtension::class)
 internal class CsvUploadServiceTest {
